New: Rudimentary plugin support
(cherry picked from commit f17bda71bf33d4c9d2f67f5ccbbf2a9378835570)

+++ src/NzbDrone.Core/History/HistoryService.cs
@@ -161,7 +161,7 @@ namespace NzbDrone.Core.History
                 history.Data.Add("Size", message.Album.Release.Size.ToString());
                 history.Data.Add("DownloadUrl", message.Album.Release.DownloadUrl);
                 history.Data.Add("Guid", message.Album.Release.Guid);
-                history.Data.Add("Protocol", ((int)message.Album.Release.DownloadProtocol).ToString());
+                history.Data.Add("Protocol", message.Album.Release.DownloadProtocol.ToString());
                 history.Data.Add("DownloadForced", (!message.Album.DownloadAllowed).ToString());
 
                 if (!message.Album.ParsedAlbumInfo.ReleaseHash.IsNullOrWhiteSpace())
